requirements.txt
streamlit==1.28.0
pandas==2.0.3
mysql-connector-python==8.2.0
plotly==5.17.0
numpy==1.24.3
streamlit-option-menu==0.3.6
